code cleanup
Former-commit-id: a3175c96ca501e6eb2d0b650817142f8120c51dc

+++ src/apps/xplorer/MSMRTCallbacks.h
@@ -112,8 +112,8 @@ public:
             }
         }
 
-        const GLint width = _cam->getViewport()->width();
-        const GLint height = _cam->getViewport()->height();
+        const GLint width = GLint( _cam->getViewport()->width() );
+        const GLint height = GLint( _cam->getViewport()->height() );
 
 #if 0
         //Make sure something is actually bound
